(refactor) : DTO 관련 코드 삭제
