--- src/main/java/racingcar/RacingCarApplication.java
@@ -1,10 +1,0 @@
-package racingcar;
-
-import racingcar.controller.GameManagerController;
-
-public class RacingCarApplication {
-
-    public static void main(String[] args) {
-        new GameManagerController().run();
-    }
-}
